Reorganize CUDA's memory header files
