+++ agency/cuda/detail/array.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <agency/detail/config.hpp>
-#include <agency/cuda/detail/memory/allocator.hpp>
+#include <agency/cuda/memory/allocator.hpp>
 #include <agency/detail/array.hpp>
 
 namespace agency
