+++ agency/cuda/detail/boxed_value.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <agency/detail/config.hpp>
-#include <agency/cuda/detail/memory/allocator.hpp>
+#include <agency/cuda/memory/allocator.hpp>
 #include <agency/detail/boxed_value.hpp>
 
 namespace agency
