+++ agency/cuda/memory.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <agency/detail/config.hpp>
-#include <agency/cuda/detail/memory/managed_allocator.hpp>
-#include <agency/cuda/detail/memory/split_allocator.hpp>
-#include <agency/cuda/detail/memory/allocator.hpp>
+#include <agency/cuda/memory/managed_allocator.hpp>
+#include <agency/cuda/memory/split_allocator.hpp>
+#include <agency/cuda/memory/allocator.hpp>
