+++ agency/cuda/memory/allocator.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <agency/detail/config.hpp>
-#include <agency/cuda/detail/memory/split_allocator.hpp>
+#include <agency/cuda/memory/split_allocator.hpp>
 #include <agency/detail/memory/caching_allocator.hpp>
 
 namespace agency
